+++ JuliaLowering/test/arrays_ir.jl
@@ -42,7 +42,7 @@ LoweringError:
 #---------------------
 LoweringError:
 [a=20, 30]
-#└──┘ ── misplaced assignment statement in `[ ... ]`
+#└──┘ ── assignment is not allowed in array expression
 
 ########################################
 # hcat syntax
@@ -86,7 +86,7 @@ T[x xs...]
 #---------------------
 LoweringError:
 [10 20 a=40]
-#     └───┘ ── misplaced assignment statement in `[ ... ]`
+#     └───┘ ── assignment is not allowed in array expression
 
 ########################################
 # vcat syntax
@@ -131,7 +131,7 @@ LoweringError:
 #---------------------
 LoweringError:
 [a=20; 30]
-#└──┘ ── misplaced assignment statement in `[ ... ]`
+#└──┘ ── assignment is not allowed in array expression
 
 ########################################
 # typed_vcat syntax